requirements.txt
Django
django-jazzmin
django-debug-toolbar
django-extensions
django-prettyjson
django-jsoneditor
djangorestframework
tzdata